Add ExportModelToPostgresCommand implementing IExternalCommand
